requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary

passlib
python-jose[cryptography]